Made Login, Register, Logout work
Login, Register are now creating access token which is being saved in the session storage. The Logout is clearing the session storage. Logged in users are unable to login and register again until they logout. Logged out users are unable to logout again unless they login or register. The navigation bar is also dynamically updating depending on whether you are logged in or not