Updated the Rent Car page mainly
The Rent Car page is now functional and it lets you rent cars if you are a logged in user, you can also cancel the rent. You can now search cars for rent based on different criteria. I added a loading gear icon on the rent car page as well and I centered it.