Updated the my profile page, edit page, added route guards
You can now update your profile picture in the my profile page. Added route guards to the logout, sell my car, my profile page using the React Router DOM library. You are unable to access the edit page if you are not logged in and if you are not the owner of the given car.